+++ examples/query-playground/src/components/QueryTab.jsx
@@ -1,4 +1,4 @@
-import { PropTypes } from 'prop-types'
+import PropTypes from 'prop-types'
 import React from 'react'
 import { QueryEditor } from './QueryEditor.jsx'
 import { QueryResults } from './QueryResults.jsx'