+++ services/data/src/index.ts
@@ -1,3 +1,7 @@
 export * from './react'
 
+export { useDataEngine, useDataQuery, useDataMutation } from './react'
+
 export { FetchError } from '@dhis2/data-engine'
+
+export type * from './types'