+++ services/data/src/react/components/CustomDataProvider.tsx
@@ -1,7 +1,7 @@
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
+import React from 'react'
 import { DataEngine, CustomDataLink } from '@dhis2/data-engine'
 import type { CustomData, CustomLinkOptions } from '@dhis2/data-engine'
-import React from 'react'
-import { QueryClient, QueryClientProvider } from 'react-query'
 import { DataContext } from '../context/DataContext'
 import { queryClientOptions as queryClientDefaults } from './DataProvider'