+++ services/data/src/types.ts
@@ -7,6 +7,8 @@ import {
     QueryResult,
 } from '@dhis2/data-engine'
 
+export type { Mutation } from './engine/types/Mutation'
+export type { Query } from './engine/types/Query'
 export interface ContextType {
     engine: DataEngine
 }
@@ -16,6 +18,7 @@ export interface ContextInput {
     apiVersion: number
 }
 
+export type ExecuteOptions = QueryExecuteOptions
 export type RefetchOptions = QueryVariables
 export type RefetchFunction<ReturnType> = (
     options?: RefetchOptions